Feature/transfer coin

+++ apps/balance/src/balance.controller.ts
@@ -13,30 +13,67 @@ import { BalanceService } from './balance.service';
 import { CryptoBalance } from '@shared/interfaces';
 import { AppError } from '@shared/AppError';
 import { LoggingService } from '@shared/logging.service';
+import { AuthService } from '@shared/auth.service';
 
 @Controller('balances')
 export class BalanceController {
   constructor(
     private readonly balanceService: BalanceService,
     private readonly logger: LoggingService,
+    private readonly authService: AuthService,
   ) {
     this.logger.setContext(BalanceController.name);
   }
 
   @Get()
-  async getAllUserBalances(): Promise<CryptoBalance[]> {
+  async getAllUserBalances(
+    @Headers('x-user-password') adminPassword: string, // Accept password in the headers
+  ): Promise<CryptoBalance[]> {
+    if (!adminPassword) {
+      throw new AppError(
+        'X-User-Password header is required',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    // Validate password
+    const isValidPassword = await this.authService.validatePassword(
+      'admin',
+      adminPassword,
+    );
+    if (!isValidPassword) {
+      throw new AppError('Invalid admin password', HttpStatus.UNAUTHORIZED);
+    }
+
     return this.balanceService.getAllBalances();
   }
 
   @Get('user')
   async getUserBalance(
     @Headers('x-user-id') userId: string,
+    @Headers('x-user-password') password: string, // Accept password in the headers
   ): Promise<CryptoBalance> {
     if (!userId) {
       throw new AppError(
         'X-User-ID header is required',
         HttpStatus.BAD_REQUEST,
       );
+    }
+
+    if (!password) {
+      throw new AppError(
+        'X-User-Password header is required',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    // Validate password
+    const isValidPassword = await this.authService.validatePassword(
+      userId,
+      password,
+    );
+    if (!isValidPassword) {
+      throw new AppError('Invalid password', HttpStatus.UNAUTHORIZED);
     }
 
     const balance = await this.balanceService.getUserBalance(userId);
@@ -49,6 +86,7 @@ export class BalanceController {
   @Get('user/value')
   async getUserBalanceValue(
     @Headers('x-user-id') userId: string,
+    @Headers('x-user-password') password: string, // Accept password in the headers
     @Query('currency') currency = 'usd',
   ): Promise<{ value: number; currency: string }> {
     if (!userId) {
@@ -56,6 +94,22 @@ export class BalanceController {
         'X-User-ID header is required',
         HttpStatus.BAD_REQUEST,
       );
+    }
+
+    if (!password) {
+      throw new AppError(
+        'X-User-Password header is required',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    // Validate password
+    const isValidPassword = await this.authService.validatePassword(
+      userId,
+      password,
+    );
+    if (!isValidPassword) {
+      throw new AppError('Invalid password', HttpStatus.UNAUTHORIZED);
     }
 
     const value = await this.balanceService.getUserBalanceValue(
@@ -71,8 +125,8 @@ export class BalanceController {
   @Put('user/rebalance')
   async rebalanceUser(
     @Headers('x-user-id') userId: string,
+    @Headers('x-user-password') password: string, // Accept password in the headers
     @Body() targetPercentages: Record<string, number>,
-    // ): Promise<void> {
   ): Promise<CryptoBalance> {
     if (!userId) {
       throw new AppError(
@@ -80,6 +134,22 @@ export class BalanceController {
         HttpStatus.BAD_REQUEST,
       );
     }
+    if (!password) {
+      throw new AppError(
+        'X-User-Password header is required',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    // Validate password
+    const isValidPassword = await this.authService.validatePassword(
+      userId,
+      password,
+    );
+    if (!isValidPassword) {
+      throw new AppError('Invalid password', HttpStatus.UNAUTHORIZED);
+    }
+
     if (!targetPercentages || Object.keys(targetPercentages).length === 0) {
       throw new AppError(
         'Balance data cannot be empty',
@@ -92,6 +162,7 @@ export class BalanceController {
   @Post('user/add')
   async createUser(
     @Headers('x-user-id') userId: string,
+    @Headers('x-user-password') password: string, // Accept password in the headers
   ): Promise<{ message: string }> {
     if (!userId) {
       throw new AppError(
@@ -99,14 +170,23 @@ export class BalanceController {
         HttpStatus.BAD_REQUEST,
       );
     }
+
+    if (!password) {
+      throw new AppError('Please provide a password', HttpStatus.BAD_REQUEST);
+    }
+
+    // Hash and store the password
+    await this.authService.addPassword(userId, password);
+
     this.logger.log(`user id: ${userId}`);
     await this.balanceService.createUser(userId);
-    return { message: `user ${userId} created succefully` };
+    return { message: `user ${userId} created successfully` };
   }
 
   @Put('user/update')
   async updateBalance(
     @Headers('x-user-id') userId: string,
+    @Headers('x-user-password') password: string, // Accept password in the headers
     @Body() body: { coin: string; amount: number },
   ): Promise<CryptoBalance> {
     if (!userId) {
@@ -114,6 +194,20 @@ export class BalanceController {
         'X-User-ID header is required',
         HttpStatus.BAD_REQUEST,
       );
+    }
+    if (!password) {
+      throw new AppError(
+        'X-User-Password header is required',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+    // Validate password
+    const isValidPassword = await this.authService.validatePassword(
+      userId,
+      password,
+    );
+    if (!isValidPassword) {
+      throw new AppError('Invalid password', HttpStatus.UNAUTHORIZED);
     }
     if (!body.coin || !body.amount) {
       throw new AppError(
@@ -130,9 +224,12 @@ export class BalanceController {
     return this.balanceService.updateBalance(userId, body.coin, body.amount);
   }
 
-  @Delete('user/remove')
-  async removeUser(
+  @Put('user/transfer')
+  async transferCoin(
     @Headers('x-user-id') userId: string,
+    @Headers('x-user-password') password: string,
+    @Body()
+    transferData: { targetUserId: string; coin: string; amount: number },
   ): Promise<CryptoBalance> {
     if (!userId) {
       throw new AppError(
@@ -140,6 +237,78 @@ export class BalanceController {
         HttpStatus.BAD_REQUEST,
       );
     }
+
+    if (!password) {
+      throw new AppError(
+        'X-User-Password header is required',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    if (
+      !transferData.targetUserId ||
+      !transferData.coin ||
+      transferData.amount === undefined ||
+      transferData.amount <= 0 ||
+      userId === transferData.targetUserId
+    ) {
+      throw new AppError(
+        'Target user ID, coin, or amount are missing or invalid',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    // Validate the user password
+    const isValidPassword = await this.authService.validatePassword(
+      userId,
+      password,
+    );
+    if (!isValidPassword) {
+      throw new AppError('Invalid password', HttpStatus.UNAUTHORIZED);
+    }
+
+    // Call the service to perform the transfer
+    const updatedBalance = await this.balanceService.transferCoin(
+      userId,
+      transferData.targetUserId,
+      transferData.coin,
+      transferData.amount,
+    );
+
+    if (!updatedBalance) {
+      throw new AppError('Transfer failed', HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+    return updatedBalance;
+  }
+
+  @Delete('user/remove')
+  async removeUser(
+    @Headers('x-user-id') userId: string,
+    @Headers('x-user-password') password: string, // Accept password in the headers
+  ): Promise<CryptoBalance> {
+    if (!userId) {
+      throw new AppError(
+        'X-User-ID header is required',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    if (!password) {
+      throw new AppError(
+        'X-User-Password header is required',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    // Validate password
+    const isValidPassword = await this.authService.validatePassword(
+      userId,
+      password,
+    );
+    if (!isValidPassword) {
+      throw new AppError('Invalid password', HttpStatus.UNAUTHORIZED);
+    }
+
     this.logger.log(`user id: ${userId}`);
     return this.balanceService.removeUser(userId);
   }
